fix: delete DELETE endpoint
